+++ src/main/java/tenno_mod/cards/common/Reposition_TENNO.java
@@ -1,56 +1,56 @@
 package tenno_mod.cards.common;
 
 import basemod.abstracts.CustomCard;
-import com.megacrit.cardcrawl.actions.AbstractGameAction;
+import com.megacrit.cardcrawl.actions.common.GainBlockAction;
 import com.megacrit.cardcrawl.cards.AbstractCard;
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.core.CardCrawlGame;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.localization.CardStrings;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
+import tenno_mod.actions.DrawAndReduceAction;
 import tenno_mod.patches.AbstractCardEnum;
 
-public class QuickSlash_TENNO extends CustomCard {
-  public static final String ID = "QuickSlash_TENNO";
+public class Reposition_TENNO extends CustomCard {
+  public static final String ID = "Reposition_TENNO";
   private static final CardStrings cardStrings = CardCrawlGame.languagePack.getCardStrings(ID);
   public static final String NAME = cardStrings.NAME;
   public static final String DESCRIPTION = cardStrings.DESCRIPTION;
-  public static final String IMG_PATH = "img/cards/QuickSlash.png";
-  private static final int COST = 0;
-  private static final int ATTACK_DMG = 2;
+  public static final String IMG_PATH = "img/cards/Reposition.png";
+  private static final int COST = 1;
+  private static final int BLOCK_AMT = 7;
+  private static final int UPGRADE_PLUS_BLOCK = 3;
 
-  private static final int UPGRADE_PLUS_DMG = 2;
-
-  public QuickSlash_TENNO() {
+  public Reposition_TENNO() {
     super(
         ID,
         NAME,
         IMG_PATH,
         COST,
         DESCRIPTION,
-        CardType.ATTACK,
+        CardType.SKILL,
         AbstractCardEnum.TENNO_COLOR,
         CardRarity.COMMON,
-        CardTarget.ENEMY
+        CardTarget.SELF
     );
-    this.baseDamage = ATTACK_DMG;
+    this.baseBlock = BLOCK_AMT;
   }
 
   public void use(AbstractPlayer p, AbstractMonster m) {
-    AbstractDungeon.actionManager.addToBottom(new com.megacrit.cardcrawl.actions.common.DamageAction(m,
-        new com.megacrit.cardcrawl.cards.DamageInfo(p, this.damage, this.damageTypeForTurn),
-        AbstractGameAction.AttackEffect.SLASH_HORIZONTAL));
-    AbstractDungeon.actionManager.addToBottom(new com.megacrit.cardcrawl.actions.common.DrawCardAction(p, 1));
+    AbstractDungeon.actionManager.addToBottom(
+        new GainBlockAction(p, p, this.block)
+    );
+    AbstractDungeon.actionManager.addToBottom(new DrawAndReduceAction(p, CardType.ATTACK));
   }
 
   public AbstractCard makeCopy() {
-    return new QuickSlash_TENNO();
+    return new Reposition_TENNO();
   }
 
   public void upgrade() {
     if (!this.upgraded) {
       upgradeName();
-      upgradeDamage(UPGRADE_PLUS_DMG);
+      upgradeBlock(UPGRADE_PLUS_BLOCK);
     }
   }
 }