+++ src/main/java/tenno_mod/cards/rare/EmbraceTheVoid_TENNO.java
@@ -19,8 +19,8 @@ public class EmbraceTheVoid_TENNO extends CustomCard {
   public static final String NAME = cardStrings.NAME;
   public static final String DESCRIPTION = cardStrings.DESCRIPTION;
   public static final String IMG_PATH = "img/cards/EmbraceTheVoid.png";
-  private static final int COST = 1;
-  private static final int MAGIC_NUMBER = 1;
+  private static final int COST = 3;
+  private static final int MAGIC_NUMBER = 2;
   private static final int UPG_MAGIC_NUMBER = 1;
 
   public EmbraceTheVoid_TENNO() {