+++ src/main/java/tenno_mod/cards/uncommon/DrinkTheKuva_TENNO.java
@@ -1,13 +1,18 @@
 package tenno_mod.cards.uncommon;
 
 import basemod.abstracts.CustomCard;
+import com.megacrit.cardcrawl.actions.animations.VFXAction;
+import com.megacrit.cardcrawl.actions.utility.SFXAction;
 import com.megacrit.cardcrawl.cards.AbstractCard;
 import com.megacrit.cardcrawl.characters.AbstractPlayer;
 import com.megacrit.cardcrawl.core.CardCrawlGame;
 import com.megacrit.cardcrawl.dungeons.AbstractDungeon;
 import com.megacrit.cardcrawl.localization.CardStrings;
 import com.megacrit.cardcrawl.monsters.AbstractMonster;
+import com.megacrit.cardcrawl.ui.panels.EnergyPanel;
+import com.megacrit.cardcrawl.vfx.combat.MindblastEffect;
 import tenno_mod.actions.DrinkKuvaAction;
+import tenno_mod.actions.TransferrenceBeamAction;
 import tenno_mod.patches.AbstractCardEnum;
 
 public class DrinkTheKuva_TENNO extends CustomCard {
@@ -17,9 +22,7 @@ public class DrinkTheKuva_TENNO extends CustomCard {
   public static final String DESCRIPTION = cardStrings.DESCRIPTION;
   public static final String DESCRIPTION_UPG = cardStrings.UPGRADE_DESCRIPTION;
   public static final String IMG_PATH = "img/cards/DrinkTheKuva.png";
-  private static final int COST = 1;
-  private static final int MAGIC_NUMBER = 6;
-  private static final int UPG_MAGIC_NUMBER = 2;
+  private static final int COST = -1;
 
   public DrinkTheKuva_TENNO() {
     super(
@@ -33,13 +36,19 @@ public class DrinkTheKuva_TENNO extends CustomCard {
         CardRarity.UNCOMMON,
         CardTarget.SELF
     );
-    this.exhaust = true;
-    this.magicNumber = this.baseMagicNumber = MAGIC_NUMBER;
   }
 
   public void use(AbstractPlayer p, AbstractMonster m) {
+    if (this.energyOnUse < EnergyPanel.totalCount) {
+      this.energyOnUse = EnergyPanel.totalCount;
+    }
+
+    int amount = energyOnUse;
+    if (this.upgraded) {
+      amount ++;
+    }
     AbstractDungeon.actionManager.addToBottom(
-        new DrinkKuvaAction(p, this.magicNumber));
+        new DrinkKuvaAction(p, amount, p, this.freeToPlayOnce));
   }
 
   public AbstractCard makeCopy() {
@@ -49,7 +58,8 @@ public class DrinkTheKuva_TENNO extends CustomCard {
   public void upgrade() {
     if (!this.upgraded) {
       upgradeName();
-      upgradeMagicNumber(UPG_MAGIC_NUMBER);
+      this.rawDescription = DESCRIPTION_UPG;
+      initializeDescription();
     }
   }
 }